+++ bin/tests.sh
@@ -18,16 +18,15 @@ mkdir -p ../onlysnarf/logs
 # schedule
 # post
 # message
-show="show-window"
 show="blank"
+show="show"
 
-verbose="verbose"
-verbose="verbose -verboser"
-verbose="verbose -verboser -verbosest"
+verbose="blank"
+verbose="verbose -verbose -verbose"
 
 declare -a testing
 testing=("schedule")
-testing=("discount" "expiration" "message" "poll" "post" "schedule" "upload")
+testing=("discount" "expiration" "message" "poll" "post" "schedule")
 
 echo "##################################################" >> ../onlysnarf/logs/tests.txt 2>&1
 echo "################## Start #########################" >> ../onlysnarf/logs/tests.txt 2>&1
@@ -36,8 +35,8 @@ echo "##################################################" >> ../onlysnarf/logs/t
 function testes() {
 	echo "Running: "$test
 
-	../onlysnarf/tests/$test.sh $show $verbose
-	# ../onlysnarf/tests/$test.sh $show >> ../onlysnarf/logs/tests.txt 2>&1
+	bin/tests/$test.sh $show $verbose
+	# ../onlysnarf/$test.sh $show >> ../onlysnarf/logs/tests.txt 2>&1
 
 	echo "##################################################" >> ../onlysnarf/logs/tests.txt 2>&1
 	echo "##################################################" >> ../onlysnarf/logs/tests.txt 2>&1