+++ onlysnarfer/bin/tests/discount.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+
+# Discount
+echo "[*] Discount - Recent"
+sudo onlysnarfpy \
+-debug \
+-$2 \
+-action "discount" \
+-user "recent" \
+-amount 40 \
+-$1 \
+-months 3 \
+-prefer-local