+++ CHANGELOG.md
@@ -300,9 +300,16 @@
   - mostly functional
   
 ----------------------------------------
+  **2.16.10** debugging pre 2.17.0
+  - upload -> post
+  - Message, Folder (unused), File, Google_File, Video, Image classes
+  - ffmpeg.py
+  - login function cleaned up / spawn_browser
+  - settings -> argsparse
+  - menu cleaned up
+  - file system selection cleaned up
 
   **2.17.0**
-  - ffmpeg / file
 
   **2.16.**
   - tested: settings get